// src/features/users/types/get-all-users.types.ts
export interface UserResponseDTO {
    id: number;
    email: string;
    username: string;
    password: string;
    name: {
        firstname: string;
        lastname: string;
    };
    address: {
        city: string;
        street: string;
        number: number;
        zipcode: string;
        geolocation: {
            lat: string;
            long: string;
        };
    };
    phone: string;
}

export type UsersListResponseDTO = UserResponseDTO[];